chore: Setup quality check CI/CD workflow
This CI/CD workflow checks the quality of a given push or pull request
by running static analysis checks, tests, and REUSE 3.3 compliance. It
produces basic artifacts that contain the current state of a build for
Linux, Windows, and MacOS. This is done so I can easily download these
artifacts to toy with them locally in case any of them fail so I can
more easily pin-point the issue(s) stated by the CI/CD runner.

Signed-off-by: Jason Pena <jasonpena@awkless.com>