chore(actions): Reduce lint strictness of quality checks
The cppcheck application introduces false positives into the build with
redundant warnings that cannot be shut up as easily compared to
clang-tidy. Thus, I decided to not use cppcheck in the quality check
action.

I may still use it as an alternative to clang-tidy, so I will keep
cppcheck in the build system as an optional feature.

Signed-off-by: Jason Pena <jasonpena@awkless.com>