feat(memory): Add `RegisterPair`
I improved the code of `ByteRegister` to `Register`, which now uses
template to determine the type instead of limiting the API to a byte.
Now I can define the 8-bit and 16-bit registers using the same API.

I also added the `RegisterPair` API to construct register pairs from
`Register` API. Now I can define and use register definitions in a
declarative manner. I added in a test case to make sure that the bit
manipulation logic in `RegisterPair` works correctly.

I created a new `soc` library target, which contains the SoC API for the
emulator. This was done so I could link the current code I wrote to the
unit test case code properly.

Signed-off-by: Jason Pena <jasonpena@awkless.com>