refactor: Move current code into `cocoboy::soc` namespace
I wanted to reorganize current code to use `cocoboy` as the base
namespace such that the filesystem matches this setup. I moved all code
implementing the GameBoy SoC into the `cocoboy::soc` namespace.

I also fixed the issue of the binary being the same name as the library,
now the binary is generated into a special `bin/` directory in the
`build` directory.

Signed-off-by: Jason Pena <jasonpena@awkless.com>