feat(cpu): Implement 8-bit load instructions
Finally implemented 8-bit load instructions. Still need to test some of
them, but this is a good start at hacking away at implementing the SM83
instruction set.

About a quarter of the 8-bit load instructions have been fully tested. I
plan on adding these tests in a separate commit, because this commit is
already fairly large. I got to focused on writing these instructions,
that I forgot to commit my progress along the way. I will do better to
not create such large commits in the future.

I think I will go with a more test driven development approach for this
emulator from now on.

Plus, I had to modify the `RegisterPair` type to allow decrement and
increment operations. It looks a little messy, and I need to properly
test it. However, I plan on implementing postfix decrement and increment
in the future for registers to make the opcode implementations more
compact.

Signed-off-by: Jason Pena <jasonpena@awkless.com>